Update seeing.py
Silly addition: --overwrite to be able to overwrite the input file with the seeing information

+++ seeing.py
@@ -1,4 +1,4 @@
-#! /usr/bin/env python
+su#! /usr/bin/env python
 
 # Copyright (c) 2012 Victor Terron. All rights reserved.
 # Institute of Astrophysics of Andalusia, IAA-CSIC
@@ -465,6 +465,11 @@ parser.add_option('--suffix', action = 'store', type = 'str',
                   help = "string to be appended to output images, before "
                   "the file extension, of course [default: %default]")
 
+parser.add_option('--overwrite', action = 'store_true', dest = 'overwrite',
+                  help = "if you want to add no suffix to the file (i.e. "
+                  "overwrite it) you need to activate this option. This "
+                  "is just a safety-catch to prevent mistakes. ")
+
 parser.add_option('--cores', action = 'store', type = 'int',
                   dest = 'ncores', default = defaults.ncores,
                   help = defaults.desc['ncores'])
@@ -854,8 +859,13 @@ def main(arguments = None):
             logging.debug("%s to be copied to %s" % (path, output_dir))
             history_msg1 = "Image FWHM = %.3f" % fwhms[path]
             history_msg2 = "Image elongation = %.3f" % elongs[path]
+            
+        if os.path.samefile(path, output_path) and options.overwrite:
+            # Files are the same, but overwrite active, kill'em all!
+            pass
+        else:
+            shutil.copy2(path, output_path)    
 
-        shutil.copy2(path, output_path)
         methods.owner_writable(output_path, True) # chmod u+w
         logging.debug("%s copied to %s" % (path, output_path))
         output_img = fitsimage.FITSImage(output_path)
